Add support for blackboard types (#338)
* add support for blackboard types, use std::variant instead of std::any in blackboard

* fix int64_t access

* add converter to handle get different for variant and any

* remove convertConst struct

* reenable initDefaultValues

* add support for promoting types from source to target blackboard

* fix monostate issues

* add support for accessing internal type of std::any values in blackboard

* fix some review comments

* add get with non const return value, use if constexpr for type dependent implementation of parser and get, remove yaml implementation for std::any and std::monostate

* fix timing issue

* clean up code, add comments, fix some errors

* Fix tests

* WIP

* Cleanup

* Tests working

* Fix format

* Exceptions thrown from get

* Throw more exceptions

* Add additional tests for blackboard types PR (#363)

* add bb tests for up and down conversion

* add UnknownType class to tests, add tests for get exceptions

* add tests for setting and initialization, add non constructable helper class, rename test fixture for blackboard

* add set and init test for blackboard

* use EXPECT_THROW for exception testing

* - use template for mapping tests
- use constexpr for type names
- improve exception tests, dont check exception msg
- compare UnknownTypes by using an int value, passed during initialization
- add additional tests for std::any

* add tests setWithConvertibleType and setWithoutSpecifyingType, use default constructor to create default values of base types

* activate tests for different types convertible to int64_t

* format

* remove old code

* Set yaml type for blackboard mapping tests


* Remove commented out code

* Review comments

* Use set API for mapping from src to target blackboard

This implies the same kind of type conversions are supported by mapping
& set API. This additionally avoids the need to suppress warnings in the
make variant helper & prevents conversions between type categories i.e.
bool -> double, int -> double etc. but supports conversions within the
same type category eg. int8_t -> int64_t, float -> double etc

* Add support for unsigned integer types

* Address minor review comments

* Fix format

+++ supplementary/alica_ros1/alica_tests/src/test/Expr/src/conditions/conditions.cpp
@@ -84,7 +84,7 @@ bool conditionSecondCall2FirstCall1237521027685048666(const Blackboard* input, c
     /*PROTECTED REGION ID(condition1237521027685048666) ENABLED START*/
     LockedBlackboardRO bb(*(rp->getBasicPlan()->getBlackboard()));
     auto* testWm = const_cast<alicaTests::TestWorldModel*>(dynamic_cast<const alicaTests::TestWorldModel*>(wm));
-    testWm->passedParameters["planInputKey"] = bb.get<int>("planInputKey");
+    testWm->passedParameters["planInputKey"] = bb.get<int64_t>("planInputKey");
     return false;
     /*PROTECTED REGION END*/
 }
@@ -198,7 +198,7 @@ bool conditionCounterCalled2901825906319407673(const Blackboard* input, const Ru
 {
     /*PROTECTED REGION ID(condition2901825906319407673) ENABLED START*/
     LockedBlackboardRO bb(*input);
-    return CounterClass::called == bb.get<int>("numberOfCalls");
+    return CounterClass::called == bb.get<int64_t>("numberOfCalls");
     /*PROTECTED REGION END*/
 }
 bool conditionSwitchIsNotSet3016035752801585170(const Blackboard* input, const RunningPlan* rp, const IAlicaWorldModel* wm)
@@ -238,7 +238,7 @@ bool conditionSimpleSwitchIsSet3787001793582633602(const Blackboard* input, cons
 {
     /*PROTECTED REGION ID(condition3787001793582633602) ENABLED START*/
     LockedBlackboardRO bb(*input);
-    return SimpleSwitches::isSet(bb.get<int>("idx"));
+    return SimpleSwitches::isSet(bb.get<int64_t>("idx"));
     /*PROTECTED REGION END*/
 }
 bool conditionBehaviourInSubPlan2EndTest3828316183435191952(const Blackboard* input, const RunningPlan* rp, const IAlicaWorldModel* wm)
